Rework Stat system to only work with integer stats

+++ MARDEK Engine/Assets/Scripts/MARDEK/Skill/ModifyStat.cs
@@ -12,7 +12,7 @@ namespace MARDEK.Skill
         public override void Apply(IStats user, IStats target)
         {
             var value = -valueExpresion.Evaluate(user, target);
-            target.ModifyStat(targetStat, value);
+            target.ModifyStat(targetStat, (int)value);
             Debug.Log($"Stat {targetStat} modified by {value}");
         }
     }
